domain-cert/requirements.txt
amqp==2.5.1
anyjson==0.2.5
APScheduler==3.2.0
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.9.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.1.0
celery==4.3.0
certifi==2021.10.8
cffi==1.12.3
charset-normalizer==2.0.8
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django==2.1.11
django-apscheduler==0.3.1
django-bootstrap3==11.1.0
django-celery-beat==1.5.0
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-ranged-response==0.2.0
django-redis-cache==2.0.0
django-simple-captcha==0.5.12
django-simpleui==2021.12.7
django-timezone-field==3.0
djangorestframework==3.10.2
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.16.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.4
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==7.2.0
Naked==0.1.31
packaging==21.3
passlib==1.7.1
Pillow==6.1.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.19
pycryptodome==3.11.0
PyJWT==2.3.0
pylint==2.12.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==19.0.0
pyparsing==3.0.6
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.3.8
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.2
typed-ast==1.5.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.25.3
vine==1.3.0
wrapt==1.13.3
zipp==0.6.0
